# docs/swagger.yaml
definitions:
  models.BookingObject:
    properties:
      conferenceLink:
        type: string
      createdAt:
        type: string
      id:
        type: string
      slotId:
        type: string
      status:
        type: string
      userId:
        type: string
    type: object
  models.BookingResponse:
    properties:
      booking:
        $ref: '#/definitions/models.BookingObject'
    type: object
  models.CreateBookingRequest:
    properties:
      createConferenceLink:
        type: boolean
      slotId:
        type: string
    required:
    - slotId
    type: object
...
  models.DummyLoginRequest:
    properties:
      role:
        enum:
        - admin
        - user
        type: string
    required:
    - role
    type: object
  models.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
  models.ListBookingResponse:
    properties:
      bookings:
        items:
          $ref: '#/definitions/models.BookingObject'
        type: array
    type: object
  models.ListBookingResponseWithPag:
    properties:
      bookings:
        items:
          $ref: '#/definitions/models.BookingObject'
        type: array
      pagination:
        $ref: '#/definitions/models.Pagination'
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      token:
        type: string
    type: object
  models.Pagination:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        enum:
        - admin
        - user
        type: string
    required:
    - email
    - password
    - role
    type: object
...
  models.UserObject:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
  models.UserResponse:
    properties:
      user:
        $ref: '#/definitions/models.UserObject'
    type: object
host: localhost:8080
info:
  contact: {}
  description: Сервис бронирования переговорок
  title: Room Booking Service API
  version: 1.0.0
paths:
  /bookings/{bookingId}/cancel:
    post:
      consumes:
      - application/json
      description: Отмена брони (только user, идемпотентно)
      parameters:
      - description: ID брони
        in: path
        name: bookingId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отменить бронь
      tags:
      - bookings
  /bookings/create:
    post:
      consumes:
      - application/json
      description: Создание брони на слот (только user)
      parameters:
      - description: Данные брони
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать бронь
      tags:
      - bookings
  /bookings/list:
    get:
      consumes:
      - application/json
      description: Получить список всех броней с пагинацией (только admin)
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Размер страницы
        in: query
        maximum: 100
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListBookingResponseWithPag'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Все брони
      tags:
      - bookings
  /bookings/my:
    get:
      consumes:
      - application/json
      description: Получить список броней текущего пользователя (только user)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListBookingResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Мои брони
      tags:
      - bookings
  /dummyLogin:
    post:
      consumes:
      - application/json
      description: Получить JWT токен по роли (admin/user)
      parameters:
      - description: Роль пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DummyLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Тестовый логин
      tags:
      - auth
  /login:
    post:
      consumes:
      - application/json
      description: Авторизация по email и паролю
      parameters:
      - description: Учетные данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Авторизация
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Регистрация нового пользователя
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Регистрация
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"